Move the store creation in a separate file to be able to use it in tests
